Move AudioPlayer to own package
