+++ musicdao/src/main/java/com/example/musicdao/player/AudioPlayer.kt
@@ -1,8 +1,10 @@
-package com.example.musicdao
+package com.example.musicdao.player
 
 import android.net.Uri
 import android.os.Bundle
 import android.view.View
+import com.example.musicdao.MusicBaseFragment
+import com.example.musicdao.R
 import com.google.android.exoplayer2.ExoPlayerFactory
 import com.google.android.exoplayer2.SimpleExoPlayer
 import com.google.android.exoplayer2.source.MediaSource
