+++ musicdao/src/main/java/com/example/musicdao/playlist/ReleaseFragment.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.musicdao.MusicService
 import com.example.musicdao.R
-import com.example.musicdao.AudioPlayer
+import com.example.musicdao.player.AudioPlayer
 import com.example.musicdao.ui.TipArtistDialog
 import com.example.musicdao.util.Util
 import com.example.musicdao.wallet.CryptoCurrencyConfig
